refactor: rearranged methods to appropriate files
